# ta-project.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "My Project",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/karyawans": {
      "get": {
        "summary": "Read",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "post": {
        "summary": "Create",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nama": {
                    "type": "string"
                  },
                  "jenis_kelamin": {
                    "type": "string"
                  },
                  "no_hp": {
                    "type": "string"
                  },
                  "alamat": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "divisiId": {
                    "type": "string"
                  },
                  "jabatanId": {
                    "type": "string"
                  },
                  "statusId": {
                    "type": "string"
                  }
                },
                "required": [
                  "nama",
                  "jenis_kelamin",
                  "no_hp",
                  "email",
                  "alamat",
                  "divisiId",
                  "jabatanId",
                  "statusId"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/karyawans/{id}": {
      "get": {
        "summary": "Get By Id",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "put": {
        "summary": "Update",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nama": {
                    "type": "string"
                  },
                  "no_hp": {
                    "type": "string"
                  },
                  "alamat": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "status_kepegawaian": {
                    "type": "string"
                  },
                  "divisiId": {
                    "type": "string"
                  },
                  "jabatanId": {
                    "type": "string"
                  }
                },
                "required": [
                  "nama",
                  "no_hp",
                  "email",
                  "status_kepegawaian",
                  "alamat",
                  "divisiId",
                  "jabatanId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "delete": {
        "summary": "Delete",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {}
  },
  "servers": [
    {
      "url": "https://your-server-url-here/api",
      "description": "Ganti URL ini dengan server Anda sendiri"
    }
  ],
  "security": []
}